try a factory approach
